Moved Events.Store.EventMetadata to Events.Metadata
